Split tsserver project system tests to seperate files
